Revert "[Router] 강의를 듣고 라우터 적용"
This reverts commit 0e1aa0c0afb7cbd9c3367129f8bc8e9e2fd26b22.
