--- src/AboutPage.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export default function AboutPage() {
-  return (
-    <div>
-      <h2>이 서비스에 대해서...</h2>
-      <p>이 서비스는 배움의 단계에 있는 서비스입니다.</p>
-      <p>허덕이고 있으니 구해주세요!</p>
-    </div>
-  );
-}
