--- src/AboutPage.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import { render } from '@testing-library/react';
-
-import AboutPage from './AboutPage';
-
-test('AboutPage', () => {
-  render(<AboutPage />);
-});
