+++ src/App.jsx
@@ -1,15 +1,14 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+
+import { useDispatch } from 'react-redux';
+
+import RegionsContainer from './RegionsContainer';
+import CategoriesContainer from './CategoriesContainer';
+import RestaurantsContainer from './RestaurantsContainer';
 
 import {
-  Switch,
-  Route,
-  Link,
-} from 'react-router-dom';
-
-import HomePage from './HomePage';
-import AboutPage from './AboutPage';
-import NotFoundPage from './NotFoundPage';
-import RestaurantsPage from './RestaurantsPage';
+  loadInitialData,
+} from './actions';
 
 // 0. 지역, 분류 목록을 얻기
 // 1. 지역 선택 - Regions <- API (0)
@@ -17,28 +16,17 @@ import RestaurantsPage from './RestaurantsPage';
 // 3. 식당 목록 - Restaurants <- API (with region, category) -> 1, 2 모두 완료된 경우
 
 export default function App() {
-  // const { location: { pathname } } = window;
-  //
-  // const MyComponent = {
-  //   '/': HomePage,
-  //   '/about': AboutPage,
-  //   '/restaurants': RestaurantsPage,
-  // }[pathname] || NotFoundPage;
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(loadInitialData());
+  });
 
   return (
     <div>
-      <header>
-        <h1>
-          <Link to="/">이 부분은 헤더 영역</Link>
-        </h1>
-      </header>
-      <Switch>
-        <Route exact path="/" component={HomePage} />
-        <Route path="/about" component={AboutPage} />
-        <Route path="/restaurants" component={RestaurantsPage} />
-        <Route component={NotFoundPage} />
-        {/* 위에서 부터 매칭하다가 결국 NotFound */}
-      </Switch>
+      <RegionsContainer />
+      <CategoriesContainer />
+      <RestaurantsContainer />
     </div>
   );
 }
