--- src/HomePage.test.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import { MemoryRouter } from 'react-router-dom';
-
-import { render } from '@testing-library/react';
-
-import HomePage from './HomePage';
-
-test('HomePage', () => {
-  render((
-    <MemoryRouter>
-      <HomePage />
-    </MemoryRouter>
-  ));
-});
