--- src/NotFoundPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function NotFoundPage() {
-  return (
-    <p>404 - Not Found</p>
-  );
-}
