+++ webpack.config.js
@@ -14,9 +14,4 @@ module.exports = {
   resolve: {
     extensions: ['.js', '.jsx'],
   },
-  devServer: {
-    historyApiFallback: {
-      index: 'index.html',
-    },
-  },
 };
